Break API clients into their own files
